+++ tools/codegen/Parser/LanguageTypes/Field.cpp
@@ -1,4 +1,6 @@
-#include "Precompiled.hpp"
+#include "Cursor.hpp"
+
+#include "MetaDataConfig.hpp"
 
 #include "LanguageTypes/Class.hpp"
 #include "LanguageTypes/Field.hpp"
@@ -6,8 +8,7 @@
 Field::Field(
     const Cursor &cursor,
     const Namespace &currentNamespace,
-    Class *parent
-    )
+    Class *parent)
   : LanguageType(cursor, currentNamespace)
   , m_isConst(cursor.GetType().IsConst())
   , m_parent(parent)
@@ -49,13 +50,24 @@ bool Field::isSetterAccessible(void) const
       (!m_isConst && m_accessModifier == CX_CXXPublic);
 }
 
+std::string Field::GetForceType(MetaDataManager const & metaData)
+{
+  if (metaData.HasProperty(Props::ForceCreate))
+  {
+    std::string const value = metaData.GetNativeString(Props::ForceCreate);
+    return value.empty() ? "ScType::Node" : value;
+  }
+
+  return "";
+}
+
 void Field::GenarateInitCode(std::stringstream & outCode) const
 {
   if (m_metaData.HasProperty(Props::Keynode))
   {
     GenerateResolveKeynodeCode(m_metaData.GetNativeString(Props::Keynode),
                                m_displayName,
-                               (m_metaData.HasProperty(Props::ForceCreate) ? true : false),
+                               GetForceType(m_metaData),
                                outCode);
   }
   else if (m_metaData.HasProperty(Props::Template))
@@ -69,16 +81,26 @@ void Field::GenerateTemplateBuildCode(std::string const & sysIdtf, std::string c
 {
   std::string vName = displayName + "_Addr_";
   outCode << "ScAddr " << vName << "; ";
-  GenerateResolveKeynodeCode(sysIdtf, vName, false, outCode);
+  GenerateResolveKeynodeCode(sysIdtf, vName, "", outCode);
   outCode << " if (result) { result = result && ctx.HelperBuildTemplate(" << displayName << ", " << vName << "); }";
 }
 
-void Field::GenerateResolveKeynodeCode(std::string const & sysIdtf, std::string const & displayName, bool forceCreation, std::stringstream & outCode)
+void Field::GenerateResolveKeynodeCode(
+  std::string const & sysIdtf, 
+  std::string const & displayName,
+  std::string const & forceCreation,
+  std::stringstream & outCode)
 {
-  outCode << "result = result && ctx.HelperResolveSystemIdtf(\""
-          << sysIdtf << "\", "
-          << displayName << ", "
-          << (forceCreation ? "true" : "false") << ");";
+  outCode << "  result = result && ctx.HelperResolveSystemIdtf(\""
+    << sysIdtf << "\", "
+    << displayName;
+  
+  if (!forceCreation.empty())
+  {
+    outCode << ", " << forceCreation;
+  }         
+
+  outCode << ");";
 }
 
 std::string const & Field::GetDisplayName() const